fix: address React Doctor queue warnings
